Restored teleprompter state machine files so we can see diffs
