+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/narration/teleprompter/TeleprompterState.kt
@@ -27,7 +27,10 @@ enum class TeleprompterItemState {
     RECORD_AGAIN,
     RECORD_AGAIN_ACTIVE,
     RECORD_AGAIN_PAUSED,
-    RECORD_AGAIN_DISABLED
+    RECORD_AGAIN_DISABLED,
+    PLAYING,
+    PLAYING_DISABLED,
+    PLAYING_WHILE_RECORDING_PAUSED,
 }
 
 interface TeleprompterState {
@@ -70,6 +73,8 @@ object RecordState : TeleprompterState {
         TeleprompterItemState.RECORD,
         TeleprompterItemState.RECORD_ACTIVE,
         TeleprompterItemState.RECORD_DISABLED,
+        TeleprompterItemState.PLAYING,
+        TeleprompterItemState.PLAYING_DISABLED
     )
 
     override val disabledState = RecordDisabledState
@@ -83,6 +88,8 @@ object RecordState : TeleprompterState {
             TeleprompterItemState.RECORD -> RecordState
             TeleprompterItemState.RECORD_ACTIVE -> RecordActiveState
             TeleprompterItemState.RECORD_DISABLED -> RecordDisabledState
+            TeleprompterItemState.PLAYING -> PlayingState
+            TeleprompterItemState.PLAYING_DISABLED -> PlayingDisabledState
             else -> {
                 throw IllegalStateException("State: $type tried to transition to state: $request")
             }
@@ -95,7 +102,7 @@ object RecordDisabledState : TeleprompterState {
 
     override val validStateTransitions = setOf(
         TeleprompterItemState.RECORD,
-        TeleprompterItemState.RECORD_ACTIVE
+        TeleprompterItemState.RECORD_ACTIVE,
     )
 
     override val disabledState = RecordDisabledState
@@ -150,7 +157,8 @@ object RecordPausedState : TeleprompterState {
     override val validStateTransitions = setOf(
         TeleprompterItemState.RECORD_ACTIVE,
         TeleprompterItemState.RECORD_AGAIN,
-        TeleprompterItemState.RECORD_AGAIN_DISABLED
+        TeleprompterItemState.RECORD_AGAIN_DISABLED,
+        TeleprompterItemState.PLAYING_WHILE_RECORDING_PAUSED,
     )
 
     override val disabledState = RecordDisabledState
@@ -164,6 +172,7 @@ object RecordPausedState : TeleprompterState {
             TeleprompterItemState.RECORD_ACTIVE -> RecordActiveState
             TeleprompterItemState.RECORD_AGAIN -> RecordAgainState
             TeleprompterItemState.RECORD_AGAIN_DISABLED -> RecordAgainDisabledState
+            TeleprompterItemState.PLAYING_WHILE_RECORDING_PAUSED -> PlayingWhileRecordingPausedState
             else -> {
                 throw IllegalStateException("State: $type tried to transition to state: $request")
             }
@@ -171,12 +180,95 @@ object RecordPausedState : TeleprompterState {
     }
 }
 
+
+object PlayingWhileRecordingPausedState : TeleprompterState {
+    override val type = TeleprompterItemState.PLAYING_WHILE_RECORDING_PAUSED
+
+    override val validStateTransitions = setOf(
+        TeleprompterItemState.RECORDING_PAUSED,
+    )
+
+    override val disabledState = PlayingDisabledState
+
+    override fun changeState(request: TeleprompterItemState): TeleprompterState {
+        if (request !in validStateTransitions) {
+            throw IllegalStateException("State: ${type} tried to transition to state: $request")
+        }
+
+        return when (request) {
+            TeleprompterItemState.RECORDING_PAUSED -> RecordPausedState
+            else -> {
+                throw IllegalStateException("State: ${type} tried to transition to state: $request")
+            }
+        }
+    }
+}
+
+
+object PlayingState : TeleprompterState {
+    override val type = TeleprompterItemState.PLAYING
+
+    override val validStateTransitions = setOf(
+        TeleprompterItemState.RECORD,
+        TeleprompterItemState.RECORD_AGAIN,
+    )
+
+    override val disabledState = PlayingDisabledState
+
+    override fun changeState(request: TeleprompterItemState): TeleprompterState {
+        if (request !in validStateTransitions) {
+            throw IllegalStateException("State: ${type} tried to transition to state: $request")
+        }
+
+        return when (request) {
+            TeleprompterItemState.RECORD -> RecordState
+            TeleprompterItemState.RECORD_AGAIN -> RecordAgainState
+            else -> {
+                throw IllegalStateException("State: ${type} tried to transition to state: $request")
+            }
+        }
+    }
+}
+
+
+object PlayingDisabledState : TeleprompterState {
+    override val type = TeleprompterItemState.PLAYING_DISABLED
+
+    override val validStateTransitions = setOf(
+        TeleprompterItemState.RECORD,
+        TeleprompterItemState.RECORD_AGAIN,
+        TeleprompterItemState.RECORD_DISABLED,
+        TeleprompterItemState.RECORD_AGAIN_DISABLED,
+    )
+
+    override val disabledState = PlayingDisabledState
+
+    override fun changeState(request: TeleprompterItemState): TeleprompterState {
+        if (request !in validStateTransitions) {
+            throw IllegalStateException("State: ${type} tried to transition to state: $request")
+        }
+
+        return when (request) {
+            TeleprompterItemState.RECORD -> RecordState
+            TeleprompterItemState.RECORD_AGAIN -> RecordAgainState
+            TeleprompterItemState.RECORD_DISABLED -> RecordDisabledState
+            TeleprompterItemState.RECORD_AGAIN_DISABLED -> RecordAgainDisabledState
+            else -> {
+                throw IllegalStateException("State: ${type} tried to transition to state: $request")
+            }
+        }
+    }
+}
+
+
 object RecordAgainState : TeleprompterState {
     override val type = TeleprompterItemState.RECORD_AGAIN
 
     override val validStateTransitions = setOf(
         TeleprompterItemState.RECORD_AGAIN_ACTIVE,
         TeleprompterItemState.RECORD_AGAIN_DISABLED,
+        TeleprompterItemState.PLAYING,
+        TeleprompterItemState.PLAYING_DISABLED
     )
 
     override val disabledState = RecordAgainDisabledState
@@ -189,6 +281,8 @@ object RecordAgainState : TeleprompterState {
         return when (request) {
             TeleprompterItemState.RECORD_AGAIN_ACTIVE -> RecordAgainActiveState
             TeleprompterItemState.RECORD_AGAIN_DISABLED -> RecordAgainDisabledState
+            TeleprompterItemState.PLAYING -> PlayingState
+            TeleprompterItemState.PLAYING_DISABLED -> PlayingDisabledState
             else -> {
                 throw IllegalStateException("State: $type tried to transition to state: $request")
             }
@@ -200,7 +294,8 @@ object RecordAgainDisabledState : TeleprompterState {
     override val type = TeleprompterItemState.RECORD_AGAIN_DISABLED
 
     override val validStateTransitions = setOf(
-        TeleprompterItemState.RECORD_AGAIN
+        TeleprompterItemState.RECORD_AGAIN,
+        TeleprompterItemState.PLAYING_DISABLED
     )
 
     override val disabledState = RecordAgainDisabledState
@@ -212,6 +307,7 @@ object RecordAgainDisabledState : TeleprompterState {
 
         return when (request) {
             TeleprompterItemState.RECORD_AGAIN -> RecordAgainState
+            TeleprompterItemState.PLAYING_DISABLED -> PlayingDisabledState
             else -> {
                 throw IllegalStateException("State: $type tried to transition to state: $request")
             }
