+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/narration/teleprompter/TeleprompterStateTransition.kt
@@ -18,18 +18,8 @@
  */
 package org.wycliffeassociates.otter.common.domain.narration.teleprompter
 
-enum class TeleprompterStateTransition {
-    RECORD,
-    PAUSE_RECORDING,
-    RESUME_RECORDING,
-    NEXT,
-    RECORD_AGAIN,
-    RESUME_RECORD_AGAIN,
-    PAUSE_RECORD_AGAIN,
-    SAVE
-}
 
-object RecordAction {
+object RecordVerseAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -41,28 +31,17 @@ object RecordAction {
 
         contexts[index].changeState(TeleprompterItemState.RECORD_ACTIVE)
 
+        // Make next item available to record
         if (index < contexts.lastIndex) {
-            var i = index + 1
-
-            while (i < contexts.size) {
-                // enable the next unrecorded node
-                if (contexts[i].state.type == TeleprompterItemState.RECORD_DISABLED) {
-                    contexts[i].changeState(TeleprompterItemState.RECORD)
-                    break
-                }
+            contexts[index + 1].changeState(TeleprompterItemState.RECORD)
+            for (i in index + 1..contexts.lastIndex) {
                 contexts[i].disable()
-                i++
-            }
-
-            // disable the remaining nodes
-            (i until contexts.size).forEach {
-                contexts[it].disable()
             }
         }
     }
 }
 
-object PauseRecordingAction {
+object PauseVerseRecordingAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -79,7 +58,7 @@ object PauseRecordingAction {
     }
 }
 
-object ResumeRecordAction {
+object ResumeVerseRecordAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -91,42 +70,31 @@ object ResumeRecordAction {
 
         contexts[index].changeState(TeleprompterItemState.RECORD_ACTIVE)
 
+        // Make next item available to record
         if (index < contexts.lastIndex) {
-            var i = index + 1
-
-            while (i < contexts.size) {
-                // enable the next unrecorded node
-                if (contexts[i].state.type == TeleprompterItemState.RECORD_DISABLED) {
-                    contexts[i].changeState(TeleprompterItemState.RECORD)
-                    break
-                }
+            contexts[index + 1].changeState(TeleprompterItemState.RECORD)
+            for (i in index + 1..contexts.lastIndex) {
                 contexts[i].disable()
-                i++
-            }
-
-            // disable the remaining nodes
-            (i until contexts.size).forEach {
-                contexts[it].disable()
             }
         }
     }
 }
 
 object NextVerseAction {
-    fun apply(contexts: MutableList<TeleprompterStateContext>, currentIndex: Int) {
-        val wasActive = contexts[currentIndex].state.type == TeleprompterItemState.RECORD_ACTIVE
+    fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
+        val wasActive = contexts[index - 1].state.type == TeleprompterItemState.RECORD_ACTIVE
 
         if (wasActive) {
-            contexts[currentIndex].changeState(TeleprompterItemState.RECORD_AGAIN_DISABLED)
-            contexts.firstOrNull { it.state.type == TeleprompterItemState.RECORD_DISABLED }?.changeState(TeleprompterItemState.RECORD_ACTIVE)
+            contexts[index - 1].changeState(TeleprompterItemState.RECORD_AGAIN_DISABLED)
+            contexts[index].changeState(TeleprompterItemState.RECORD_ACTIVE)
         } else {
-            contexts[currentIndex].changeState(TeleprompterItemState.RECORD_AGAIN)
-            contexts.firstOrNull { it.state.type == TeleprompterItemState.RECORD_DISABLED }?.changeState(TeleprompterItemState.RECORD)
+            contexts[index - 1].changeState(TeleprompterItemState.RECORD_AGAIN)
+            contexts[index].changeState(TeleprompterItemState.RECORD)
         }
     }
 }
 
-object RecordAgainAction {
+object RecordVerseAgainAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -146,13 +114,13 @@ object RecordAgainAction {
     }
 }
 
-object PauseRecordAgainAction {
+object PauseRecordVerseAgainAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         contexts[index].changeState(TeleprompterItemState.RECORD_AGAIN_PAUSED)
     }
 }
 
-object ResumeRecordAgainAction {
+object ResumeRecordVerseAgainAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -173,7 +141,7 @@ object ResumeRecordAgainAction {
     }
 }
 
-object SaveRecordingAction {
+object SaveVerseRecordingAction {
     fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
         if (index !in contexts.indices) return
 
@@ -182,6 +150,8 @@ object SaveRecordingAction {
                 contexts[i].restore()
                 if (contexts[i].state.type == TeleprompterItemState.RECORD_AGAIN_DISABLED) {
                     contexts[i].changeState(TeleprompterItemState.RECORD_AGAIN)
+                } else if (contexts[i].state.type == TeleprompterItemState.RECORD_DISABLED) {
+                    contexts[i].changeState(TeleprompterItemState.RECORD)
                 }
             }
         }
@@ -191,6 +161,54 @@ object SaveRecordingAction {
         if (index < contexts.lastIndex) {
             for (i in index + 1..contexts.lastIndex) {
                 contexts[i].restore()
+            }
+        }
+    }
+}
+
+
+object PlayVerseAction {
+    fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
+        if (index !in contexts.indices) return
+
+        val isVerseBeingRecorded = contexts[index].state == RecordPausedState
+
+        if (isVerseBeingRecorded) {
+            contexts[index].changeState(TeleprompterItemState.PLAYING_WHILE_RECORDING_PAUSED)
+
+            contexts.forEachIndexed { i, verseContext ->
+                if (i != index) {
+                    contexts[i].disable()
+                }
+            }
+        } else {
+            contexts[index].changeState(TeleprompterItemState.PLAYING)
+
+            contexts.forEachIndexed { i, verseContext ->
+                if (i != index) {
+                    contexts[i].disable()
+                }
+            }
+        }
+    }
+}
+
+
+object PauseVersePlaybackAction {
+    fun apply(contexts: MutableList<TeleprompterStateContext>, index: Int) {
+        if (index !in contexts.indices) return
+
+        val isVerseBeingRecorded = contexts[index].state == PlayingWhileRecordingPausedState
+
+        if (isVerseBeingRecorded) {
+            contexts[index].changeState(TeleprompterItemState.RECORDING_PAUSED)
+        } else {
+            contexts[index].changeState(TeleprompterItemState.RECORD_AGAIN)
+
+            contexts.forEachIndexed { i, verseContext ->
+                if (i != index) {
+                    contexts[i].restore()
+                }
             }
         }
     }
